株じゃないポートフォリオがあったときに無視する設定にバグがあったため修正

+++ entrypoint.sh
@@ -41,9 +41,11 @@ kill_process () {
     port=$1
     headerSample="PID"
     pid=$(lsof -i :$port | grep -v $headerSample | head -n 1 | awk '{print $2}')
-    execCmd="kill -9 $pid"
-    echo "$execCmd"
-    bash -c "$execCmd"
+    if [ "$pid" <> "" ]; then
+        execCmd="kill -9 $pid"
+        echo "$execCmd"
+        bash -c "$execCmd"
+    fi
 }
 
 RAILS_ROOT="${CONTAINER_ROOT}/stock_rails"
